video_making/requirements.txt
moviepy==1.0.3
python-dotenv==1.0.0
requests==2.31.0
pysrt==1.1.2